requirement.txt
absl-py==2.1.0
aiohttp @ file:///C:/b/abs_c4zmy2l696/croot/aiohttp_1670009573673/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
altgraph==0.17.4
ansi2html==1.8.0
astor==0.8.1
async-timeout @ file:///C:/b/abs_43ozhz2a8g/croots/recipe/async-timeout_1664876362767/work
asynctest==0.13.0
attrs @ file:///C:/b/abs_09s3y775ra/croot/attrs_1668696195628/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2023.5.7
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
chardet==3.0.4
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///C:/ci/click_1646038601470/work
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
comm==0.1.4
ConfigArgParse==1.7
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
cycler==0.11.0
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy @ file:///D:/bld/debugpy_1660619187469/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dlib==19.24.6
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
face-recognition==1.3.0
face-recognition-models==0.3.0
fastjsonschema==2.18.1
Flask==2.2.5
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
frozenlist @ file:///C:/b/abs_2bb5uzghsi/croot/frozenlist_1670004511812/work
future==1.0.0
gast==0.6.0
google-auth @ file:///opt/conda/conda-bld/google-auth_1646735974934/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
graphviz==0.8.4
grpcio==1.62.3
h5py==3.8.0
idna==2.6
imageio==2.30.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel @ file:///D:/bld/ipykernel_1666723449132/work
ipython @ file:///D:/bld/ipython_1651240762678/work
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1673615989977/work
jupyter_core @ file:///D:/bld/jupyter_core_1658332454038/work
jupyterlab-widgets==3.0.9
Keras-Applications==1.0.8
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver==1.4.5
llvmlite==0.39.1
lpips==0.1.4
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
menpo==0.11.1
meson==1.5.1
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
mpmath==1.3.0
multidict @ file:///C:/b/abs_6cx_8w3cv2/croot/multidict_1665674238352/work
mxnet==1.7.0.post2
nbformat==5.7.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==2.6.3
ninja==1.11.1.1
numba==0.56.4
numpy==1.21.6
numpy-stl==3.1.2
oauthlib @ file:///C:/b/abs_2eoymqc2ow/croot/oauthlib_1665490906043/work
open3d==0.17.0
opencv-python==4.10.0.84
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging==24.0
pandas==1.3.5
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pefile==2023.2.7
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
plotly==5.17.0
plyfile==0.9
pooch==1.8.2
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1688565951714/work
protobuf==3.20.3
psutil @ file:///D:/bld/psutil_1653089401496/work
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1681904169130/work
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2024.0
PyJWT @ file:///C:/ci/pyjwt_1657529476747/work
pyntcloud==0.3.1
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pyparsing==3.1.4
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyrsistent==0.19.3
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.9.0.post0
python-utils==3.5.2
pytz==2024.2
pyvista==0.38.6
PyWavelets==1.3.0
pywin32==303
pywin32-ctypes==0.2.2
pyzmq @ file:///D:/bld/pyzmq_1656183658660/work
requests==2.18.4
requests-oauthlib==1.3.0
retrying==1.3.4
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-image==0.18.3
scikit-learn==1.0.2
scikit-metrics==0.1.0
scipy==1.7.3
scooby==0.7.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
sympy==1.10.1
tenacity==8.2.3
tensorboard==1.14.0
tensorboard-data-server @ file:///C:/b/abs_2fhvpo862s/croot/tensorboard-data-server_1670853600144/work/tensorboard_data_server-0.6.1-py3-none-any.whl
tensorboard-plugin-wit==1.6.0.post3
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.6.0+cpu
torchaudio==0.8.1
torchvision==0.9.1
tornado @ file:///D:/bld/tornado_1656937938087/work
tqdm==4.66.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
typing_extensions==4.7.1
urllib3==1.22
vtk==9.2.5
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
Werkzeug==2.2.3
widgetsnbextension==4.0.9
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wrapt==1.16.0
yarl @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_e5nlunspj6/croots/recipe/yarl_1661437086516/work
zipp==3.15.0
